Initial work before major refactor.
